Fix locking problems by switching to fcntl lockfiles.
With the locking mechanism introduced in 1.4.20, a high rate of
contention leads to significant delay due to the sleep used to poll the
bind and potential starvation.  Even with low contention, the large
delay introduced can cause problems for existing scripts.  It is also
insecure as a normal user can hold the lock.  This method uses a fcntl
lockfile in /var/run instead to fix these issues.