Reverse the meaning of the -w argument to be --no-wait rather than --wait.
With 1.4.20, getting similar results from iptables to previous versions
required adding -w to all iptables calls or iptables could return
failure unexpectedly which caused problems with scripts.  This patch
reverses the meaning of -w such that by default iptables will simply
work as expected by locking as necessary.  If locking is undesirable,
-w can be used to disable it.  The result is that this version of
iptables can work as a drop-in replacement for iptables prior to 1.4.20
while the 1.4.20 version could not.